backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
pytest==7.4.3
pytest-asyncio==0.21.1
